allow metrictank to run as graphite 1.0 cluster node
Allow metrictank to sit behind a stock graphite-web 1.0.1 deployment
as a cluster_server.  This allows metrictank to be used without the
need of using a custom finder plugin.
To support this feature the following changes have been made
- support pickle format for metrics/find calls
- support pickle format for render calls

Docs and docker Envs have also been updated to reflect the change
from using graphite-api to using graphite-web